+++ src/game/battle.rs
@@ -1,21 +1,11 @@
 use super::Game;
-use crate::character::{Character, Dead, StatusEffect};
+use crate::character::{AttackType, Character, Dead};
 use crate::event::Event;
 use crate::item::key::Key;
-use crate::randomizer::Randomizer;
-
-/// Outcome of an attack attempt.
-/// This affects primarily how the attack is displayed.
-pub enum AttackType {
-    Regular,
-    Critical,
-    Effect(StatusEffect),
-    Miss,
-}
 
 /// Run a turn-based combat between the game's player and the given enemy.
 /// Return Ok(xp gained) if the player wins, or Err(()) if it loses.
-pub fn run(game: &mut Game, enemy: &mut Character, random: &dyn Randomizer) -> Result<i32, Dead> {
+pub fn run(game: &mut Game, enemy: &mut Character) -> Result<i32, Dead> {
     // These accumulators get increased based on the characters speed:
     // the faster will get more frequent turns.
     // This could be generalized to player vs enemy parties
@@ -28,15 +18,26 @@ pub fn run(game: &mut Game, enemy: &mut Character, random: &dyn Randomizer) -> R
 
         if pl_accum >= en_accum {
             if !autopotion(game, enemy) && !autoether(game, enemy) {
-                let new_xp = player_attack(game, enemy, random);
+                let new_xp = player_attack(game, enemy);
                 xp += new_xp;
             }
 
             game.apply_status_effects()?;
             pl_accum = -1;
         } else {
-            enemy_attack(game, enemy, random)?;
-            // TODO enemy receive status effect
+            enemy_attack(game, enemy)?;
+
+            // some duplication with game mod
+            let (hp, mp) = enemy.apply_status_effects().unwrap_or_default();
+            Event::emit(
+                game,
+                Event::StatusEffect {
+                    enemy: Some(enemy),
+                    hp,
+                    mp,
+                },
+            );
+
             en_accum = -1;
         }
     }
@@ -44,11 +45,17 @@ pub fn run(game: &mut Game, enemy: &mut Character, random: &dyn Randomizer) -> R
     Ok(xp)
 }
 
+// FIXME remove duplication between these two functions,
+// move char-specific logic to char module
 /// Attack enemy, returning the gained experience
-fn player_attack(game: &mut Game, enemy: &mut Character, random: &dyn Randomizer) -> i32 {
-    let (attack_type, damage, mp_cost, new_xp) = generate_attack(&game.player, enemy, random);
-    enemy.update_hp(-damage).unwrap_or_default();
+fn player_attack(game: &mut Game, enemy: &mut Character) -> i32 {
+    let (attack_type, damage, mp_cost, new_xp) = game.player.generate_attack(enemy);
     game.player.update_mp(-mp_cost);
+    enemy.update_hp(-damage).unwrap_or_default();
+
+    if let AttackType::Effect(status) = attack_type {
+        enemy.status_effect = Some(status);
+    }
 
     Event::emit(
         game,
@@ -63,14 +70,10 @@ fn player_attack(game: &mut Game, enemy: &mut Character, random: &dyn Randomizer
 }
 
 /// Attack player, returning Err(Dead) if the player dies.
-fn enemy_attack(
-    game: &mut Game,
-    enemy: &mut Character,
-    random: &dyn Randomizer,
-) -> Result<(), Dead> {
-    let (attack_type, damage, mp_cost, _xp) = generate_attack(enemy, &game.player, random);
-    let result = game.player.update_hp(-damage).map(|_| ());
+fn enemy_attack(game: &mut Game, enemy: &mut Character) -> Result<(), Dead> {
+    let (attack_type, damage, mp_cost, _xp) = enemy.generate_attack(&game.player);
     enemy.update_mp(-mp_cost);
+    let result = game.player.update_hp(-damage).map(|_| ());
 
     if let AttackType::Effect(status) = attack_type {
         game.player.status_effect = Some(status);
@@ -87,34 +90,6 @@ fn enemy_attack(
     result
 }
 
-/// Return randomized attack parameters according to the character attributes.
-fn generate_attack(
-    attacker: &Character,
-    receiver: &Character,
-    random: &dyn Randomizer,
-) -> (AttackType, i32, i32, i32) {
-    let (damage, mp_cost) = attacker.damage(receiver);
-    let damage = random.damage(damage);
-    let xp = attacker.xp_gained(receiver, damage);
-
-    let attack_type = random.attack_type(
-        attacker.inflicted_status_effect(),
-        attacker.speed(),
-        receiver.speed(),
-    );
-
-    match attack_type {
-        AttackType::Miss => (attack_type, 0, mp_cost, 0),
-        AttackType::Regular => (AttackType::Regular, damage, mp_cost, xp),
-        AttackType::Critical => (attack_type, damage * 2, mp_cost, xp),
-        AttackType::Effect(status) if Some(status) != receiver.status_effect => {
-            (attack_type, damage, mp_cost, xp)
-        }
-        // don't double-inflict if already has the same status
-        AttackType::Effect(_) => (AttackType::Regular, damage, mp_cost, xp),
-    }
-}
-
 /// If the player is low on hp and has a potion available use it
 /// instead of attacking in the current turn.
 fn autopotion(game: &mut Game, enemy: &Character) -> bool {
@@ -124,7 +99,7 @@ fn autopotion(game: &mut Game, enemy: &Character) -> bool {
 
     // If there's a good chance of winning the battle on the next attack,
     // don't use the potion.
-    let (potential_damage, _mp_cost) = game.player.damage(enemy);
+    let (_, potential_damage, _, _) = game.player.generate_attack(enemy);
     if potential_damage >= enemy.current_hp {
         return false;
     }
@@ -139,7 +114,7 @@ fn autoether(game: &mut Game, enemy: &Character) -> bool {
 
     // If there's a good chance of winning the battle on the next attack,
     // don't use the ether.
-    let (potential_damage, _mp_cost) = game.player.damage(enemy);
+    let (_, potential_damage, _, _) = game.player.generate_attack(enemy);
     if potential_damage >= enemy.current_hp {
         return false;
     }
@@ -152,7 +127,6 @@ mod tests {
     use super::*;
     use crate::character;
     use crate::character::class;
-    use crate::randomizer::random;
 
     #[test]
     fn won() {
@@ -232,17 +206,17 @@ mod tests {
         game.player = character::Character::new(player_class, 1);
 
         // mage -mp with enough mp
-        player_attack(&mut game, &mut enemy, &random());
+        player_attack(&mut game, &mut enemy);
         assert_eq!(7, game.player.current_mp);
         assert_eq!(70, enemy.current_hp);
 
-        player_attack(&mut game, &mut enemy, &random());
-        player_attack(&mut game, &mut enemy, &random());
+        player_attack(&mut game, &mut enemy);
+        player_attack(&mut game, &mut enemy);
         assert_eq!(1, game.player.current_mp);
         assert_eq!(10, enemy.current_hp);
 
         // mage -mp=0 without enough mp
-        player_attack(&mut game, &mut enemy, &random());
+        player_attack(&mut game, &mut enemy);
         assert_eq!(1, game.player.current_mp);
         assert_eq!(7, enemy.current_hp);
     }